requirements.txt
black
mypy>=1.18.2,<2
openai==2.8.1
pydantic==2.12.4
pydantic-settings==2.12.0
python-telegram-bot[all]==22.5
